[Bronze V] Title: A/B, Time: 128 ms, Memory: 14360 KB -BaekjoonHub

+++ Java/백준/Bronze/1008. A／B/A／B.java
@@ -8,8 +8,8 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        double A = Integer.parseInt(st.nextToken());
-        double B = Integer.parseInt(st.nextToken());
+        double A = Double.parseDouble(st.nextToken());
+        double B = Double.parseDouble(st.nextToken());
         double div = A / B;
 
         System.out.println(div);
